base: Install proper packages and configuration for sound
Not really Debian or Arch based, but for my laptop. Apparently, latest
kernel version changed something in snd_hda_intel, and the dsp_driver=1
option must be enabled. Quoting from
https://bugzilla.kernel.org/show_bug.cgi?id=205959:

    The dmic_detect parameter was removed from snd_hda_intel and
    replaced by "snd-intel-dspcfg.dsp_driver=1" kernel parameter in the
    latest kernel code.

        echo "options snd-intel-dspcfg dsp_driver=1" > /etc/modprobe.d/alsa.conf

Without this, and alsa packages, I get no sound. To be fair I don't know
whether all the packages are required, nor if the modprobe config is
required, but all of them together gets me sound, that's enough for me.